Fix API: add CORS, validate PORT, split app from server
