+++ apps/api/src/index.ts
@@ -1,27 +1,7 @@
-import express from "express";
+import app from "./app";
+import { parsePort } from "./config";
 
-import usersRouter from "./routes/users";
-import piRouter from "./routes/pi";
-
-const app = express();
-const port = process.env.PORT || 4000;
-
-// Configure JSON body size limit for security
-app.use(express.json({ limit: "100kb" }));
-
-// Health check endpoint with normalized response envelope
-app.get("/health", (_req, res) => {
-  res.json({
-    status: "ok",
-    data: {
-      service: "taskflow-api",
-      timestamp: new Date().toISOString()
-    }
-  });
-});
-
-app.use("/users", usersRouter);
-app.use("/pi", piRouter);
+const port = parsePort();
 
 app.listen(port, () => {
   console.log(`TaskFlow API listening on port ${port}`);
